% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-initial.R
\name{.initial_choices}
\alias{.initial_choices}
\title{Choices of Initial Configurations}
\usage{
.initial_choices(x)
}
\arguments{
\item{x}{Data set identifier as a character scalar.}
}
\value{
A named character vector of choices for initial states of the
\code{\link[=iSEE]{iSEE()}} app for that data set.
}
\description{
\code{.initial_choices()} identifies choices of predefined initial app states
available for a given data set, and returns them as a named character vector
meant to populate a Shiny \code{\link[=selectizeInput]{selectizeInput()}}.
}
\details{
A default option is automatically prepended to the choices for all data sets.
That default option launches an initial state that includes one panel of
each type compatible with the information present in the data set.
}
\examples{
.initial_choices("dummy_id")
}
