% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observers.R
\name{.create_observers}
\alias{.create_observers}
\alias{.create_launch_observers}
\title{Observers for iSEEindex}
\usage{
.create_observers(input, session, pObjects, rObjects)

.create_launch_observers(FUN, bfc, input, session, pObjects)
}
\arguments{
\item{input}{The Shiny input object from the server function.}

\item{session}{The Shiny session object from the server function.}

\item{pObjects}{An environment containing global parameters generated in the landing page.}

\item{rObjects}{A reactive list of values generated in the landing page.}

\item{FUN}{A function to initialize the \pkg{iSEE} observer
architecture. Refer to \code{\link[iSEE:createLandingPage]{iSEE::createLandingPage()}} for more details.}

\item{bfc}{An \code{\link[=BiocFileCache]{BiocFileCache()}} object.
landing page.}
}
\value{
Those functions create observers in the server function in which they are called.
In all cases, a \code{NULL} value is invisibly returned.
}
\description{
\code{.create_observers()} initialises observers for the \pkg{iSEEindex} landing page.
}
\details{
\code{.create_launch_observers()} initialises observers for launching the \pkg{iSEE} main app.

Currently, the \code{.create_observers()} function expects a column named \code{uri}
in the metadata table of available data sets, which it uses as the identifier
for each data set.
}
