% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landing_page.R
\name{.create_persistent_objects}
\alias{.create_persistent_objects}
\title{Create persistent objects}
\usage{
.create_persistent_objects(datasets_table)
}
\arguments{
\item{datasets_table}{A \code{data.frame} of metadata for all available data sets.}
}
\value{
An environment containing several global variables for use throughout the application.
}
\description{
Create persistent objects
}
