% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datasets.R
\name{.datasets_available}
\alias{.datasets_available}
\title{Available Data Sets}
\usage{
.datasets_available(FUN)
}
\arguments{
\item{FUN}{A function.}
}
\value{
\code{.datasets_available()} returns a \code{data.frame} of metadata for available data sets.
}
\description{
Import and process metadata for available data sets.
}
\examples{
## Setup ----

dataset_fun <- function() {
    out <- read.csv(system.file(package = "iSEEindex", "datasets.csv"))
}

## Usage ----

iSEEindex:::.datasets_available(dataset_fun)

}
