% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landing_page.R
\name{.landing_page}
\alias{.landing_page}
\title{Landing page function}
\usage{
.landing_page(bfc, FUN.datasets)
}
\arguments{
\item{bfc}{A \code{\link[=BiocFileCache]{BiocFileCache()}} object.}

\item{FUN.datasets}{A function that returns a \code{data.frame} of metadata for available data sets.
See \code{\link[=.datasets_available]{.datasets_available()}}.}
}
\value{
A \code{function} that defines UI elements and observers for the
landing page of the app.
}
\description{
Landing page function
}
