% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEindex.R
\name{.launch_isee}
\alias{.launch_isee}
\title{Prepare and Launch the Main App.}
\usage{
.launch_isee(FUN, bfc, session, pObjects)
}
\arguments{
\item{FUN}{A function to initialize the \pkg{iSEE} observer
architecture. Refer to \code{\link[iSEE:createLandingPage]{iSEE::createLandingPage()}} for more details.}

\item{bfc}{An \code{\link[=BiocFileCache]{BiocFileCache()}} object.}

\item{session}{The Shiny session object from the server function.}

\item{pObjects}{An environment containing global parameters generated in the
landing page.}
}
\value{
A \code{NULL} value is invisibly returned.
}
\description{
Prepare and Launch the Main App.
}
\details{
This function wraps steps that can be tracked and reported to the
user through a progress bar, using \code{\link[shiny:withProgress]{shiny::withProgress()}}.
}
\section{Unit testing}{

This function cannot be unit tested (yet), as \code{\link[shiny:withProgress]{shiny::withProgress()}}
requires a functional \code{ShinySession} object.

This might be revisited in the future.
}

