% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-initial.R
\name{.load_initial}
\alias{.load_initial}
\title{Load an Initial Application State}
\usage{
.load_initial(pObjects)
}
\arguments{
\item{pObjects}{An environment containing global parameters generated in the
landing page.}
}
\value{
A \code{list} of \linkS4class{Panel} representing an initial app state.
}
\description{
Load an Initial Application State
}
\details{
Currently, this function always return \code{NULL}, representing the default
initial state that includes one panel of each type compatible with the
information present in the data set.
}
