% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datasets.R
\name{.load_sce}
\alias{.load_sce}
\alias{.convert_to_sce}
\title{Load Object and Coerce to SingleCellExperiment}
\usage{
.load_sce(bfc, uri)

.convert_to_sce(x)
}
\arguments{
\item{bfc}{A}

\item{uri}{A URI as a character scalar.}

\item{x}{An object Coercible to SingleCellExperiment}
}
\value{
\code{.load_sce()} returns a \code{\link[=SingleCellExperiment]{SingleCellExperiment()}} object.

For \code{.convert_to_sce()}, a \code{\link[=SingleCellExperiment]{SingleCellExperiment()}} object.
}
\description{
Load Object and Coerce to SingleCellExperiment
}
\examples{
## Setup ----

library(BiocFileCache)
bfc <- BiocFileCache()
uri <- "https://zenodo.org/record/7186593/files/ReprocessedAllenData.rds?download=1"

## Usage ---

iSEEindex:::.load_sce(bfc, uri)

}
