% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{.render_datasets_table}
\alias{.render_datasets_table}
\title{Render Table of Available Data Sets}
\usage{
.render_datasets_table(output, pObjects, rObjects)
}
\arguments{
\item{output}{The Shiny output object from the server function.}

\item{pObjects}{An environment containing global parameters generated in the landing page.}

\item{rObjects}{A reactive list of values generated in the landing page.}
}
\value{
Adds a rendered \code{\link[DT:datatable]{DT::datatable()}} to \code{output}.
A \code{NULL} value is invisibly returned.
}
\description{
Render Table of Available Data Sets
}
