% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{.render_markdown_overview}
\alias{.render_markdown_overview}
\alias{.render_initial_overview}
\title{Render Overview of Selected Data Set}
\usage{
.render_markdown_overview(output, pObjects, rObjects)

.render_initial_overview(output, pObjects, rObjects)
}
\arguments{
\item{output}{The Shiny output object from the server function.}

\item{pObjects}{An environment containing global parameters generated in the landing page.}

\item{rObjects}{A reactive list of values generated in the landing page.}
}
\value{
Adds a rendered \code{\link[DT:datatable]{DT::datatable()}} to \code{output}.
A \code{NULL} value is invisibly returned.
}
\description{
\code{.render_markdown_overview()} renders an overview of the selected data set using Markdown.
}
\details{
\code{.render_initial_overview()} renders an overview of the selected initial configuration using Markdown.

Currently, those functions expect a column named \code{uri} in the metadata table
of available data sets, which it uses as the identifier for each data set.
}
