% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEindex.R
\name{iSEEindex}
\alias{iSEEindex}
\title{iSEEindex app}
\usage{
iSEEindex(bfc, FUN.datasets, FUN.initial = NULL)
}
\arguments{
\item{bfc}{An \code{\link[=BiocFileCache]{BiocFileCache()}} object.}

\item{FUN.datasets}{A function that returns a \code{data.frame} of metadata for
available data sets.}

\item{FUN.initial}{A function that takes a data set identifier and returns
a named character vector of R scripts that each define an initial configuration.
Default to \code{NULL}, disabling custom initial states altogether.}
}
\value{
An \code{\link[=iSEE]{iSEE()}} app with a custom landing page using a \code{\link[=BiocFileCache]{BiocFileCache()}} to cache a selection of data sets.
}
\description{
iSEEindex app
}
\examples{
library(BiocFileCache)
bfc <- BiocFileCache(cache = tempdir())

dataset_fun <- function() {
    out <- read.csv(system.file(package = "iSEEindex", "datasets.csv"))
}

initial_fun <- function(id) {
    config_table <- read.csv(system.file(package = "iSEEindex", "initial.csv"))
    config_subset_table <- subset(config_table, dataset_id == id)
    out <- config_subset_table$uri
    names(out) <- config_subset_table$label
    out
}

app <- iSEEindex(bfc, dataset_fun, initial_fun)

if (interactive()) {
  shiny::runApp(app, port = 1234)
}
}
