% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEindexResource-class.R
\name{iSEEindexLocalhostResource-class}
\alias{iSEEindexLocalhostResource-class}
\alias{precache,iSEEindexLocalhostResource-method}
\title{The iSEEindexLocalhostResource class}
\description{
The iSEEindexLocalhostResource class represents a resource accessible through
a local filepath.
A URI for this type of resource uses the prefix \dQuote{localhost://}.
}
\details{
Absolute and relative paths are both supported.

Absolute paths require an additional \code{/} (forward slash)
following the double forward slash \verb{//} separating the scheme component of the URI.

For instance:

\itemize{
\item \verb{localhost://path/to/file} refers to the relative path \code{path/to/file}
(relative to the working directory when the Shiny application is launched).
\item \verb{localhost:///path/to/file} refers to the absolute path \verb{/path/to/file}.
}
}
\section{Slot overview}{

This class inherits all slots from its parent class \linkS4class{iSEEindexResource}.
}

\section{Supported methods}{

In the following code snippets, \code{x} is an instance of a \linkS4class{iSEEindexLocalhostResource} class.
Refer to the documentation for each method for more details on the remaining arguments.

\itemize{
\item \code{\link{precache}(x, ...)} trims the \verb{localhost://} prefix, and caches a copy of the resource located at the resulting file path using \pkg{BiocFileCache}, before returning the file path to the cached file.
}
}

\examples{
new("iSEEindexLocalhostResource", uri = "localhost:///example/absolute/path")
new("iSEEindexLocalhostResource", uri = "localhost://example/relative/path")
}
\author{
Kevin Rue-Albrecht
}
