% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEindexResource-class.R
\docType{class}
\name{iSEEindexResource-class}
\alias{iSEEindexResource-class}
\alias{download,iSEEindexHttpsResource-method}
\title{The iSEEindexResource class}
\description{
The iSEEindexResource class represents a resource accessible through
an HTTPS link.
}
\section{Methods}{

The \code{download()} method returns the URI to the HTTPS link as-is,
because \pkg{BiocFileCache} is capable of using those links directly.
}

