% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEindexResource-class.R
\docType{methods}
\name{iSEEindexResource-generics}
\alias{iSEEindexResource-generics}
\alias{precache}
\title{Download and Cache Resources}
\description{
An overview of the generics for downloading and caching resources.
}
\section{Defining the parameter interface}{

\code{precache(x, ...)} potentially downloads a resource from an URI and returns the path to the local file:
\itemize{
\item \code{x}, a character scalar that represents a URI.
\item \code{...}, additional arguments passed to and from other methods.
}
}

\author{
Kevin Rue-Albrecht
}
