% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEindexResource-class.R
\name{iSEEindexRunrResource-class}
\alias{iSEEindexRunrResource-class}
\alias{precache,iSEEindexRunrResource-method}
\alias{iSEEindexRunrResource}
\title{The iSEEindexRunrResource class}
\usage{
iSEEindexRunrResource(x)
}
\arguments{
\item{x}{List of metadata. See Details.}
}
\value{
The constructor function \code{iSEEindexRunrResource()} returns an object of object of class \code{iSEEindexRunrResource}.
}
\description{
The iSEEindexRunrResource class represents an SE object, obtained directly
through an R call.
}
\details{
A URI for this type of resource uses the prefix \dQuote{runr://}.

Required metadata:

\describe{
\item{uri}{Character scalar. URI of the resource.}
}
}
\section{Supported methods}{

In the following code snippets, \code{x} is an instance of a \linkS4class{iSEEindexRunrResource} class.
Refer to the documentation for each method for more details on the remaining arguments.

\itemize{
\item \code{\link{precache}(x, bfc, id, ...)} caches the resource located at the given URI using \pkg{BiocFileCache} and returns the file path to the cached file.
}
}

\examples{
iSEEindexRunrResource(list(
  uri = "runr://HCATonsilData::HCATonsilData(assayType = 'RNA', cellType = 'epithelial')"
))
}
