% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEindexResource-class.R
\name{iSEEindexS3Resource-class}
\alias{iSEEindexS3Resource-class}
\alias{precache,iSEEindexS3Resource-method}
\title{The iSEEindexS3Resource class}
\usage{
\S4method{precache}{iSEEindexS3Resource}(x, temp_dir = tempdir(), ...)
}
\arguments{
\item{temp_dir}{Scalar character, the directory to store the downloaded file
in before it is handed over to BiocFileCache. This directory will be created
recursively if it doesn't already exist.}

\item{...}{Additional arguments passed on to the \code{\link[paws.storage:s3]{paws.storage::s3()}}
function.}
}
\description{
The iSEEindexS3Resource class represents a cloud storage resource accessible
via the
\href{https://cran.r-project.org/package=paws.storage}{paws.storage}
R package.
A URI for this type of resource uses the prefix \dQuote{s3://}.
}
\details{
The URI must correspond to an existing file in an AWS S3 compatible cloud
storage system that produces an object called \code{initial} that contains a
valid configuration for an \pkg{iSEE} app. For details about authentication,
please consult the
\href{https://github.com/paws-r/paws/blob/main/docs/credentials.md}{paws R package documentation}.

For instance:

\if{html}{\out{<div class="sourceCode">}}\preformatted{s3://your-bucket/your-prefix/index.rds)
}\if{html}{\out{</div>}}
}
\section{Slot overview}{

This class inherits all slots from its parent class \linkS4class{iSEEindexResource}.
}

\section{Supported methods}{

In the following code snippets, \code{x} is an instance of a \linkS4class{iSEEindexS3Resource} class.
Refer to the documentation for each method for more details on the remaining arguments.

\itemize{
\item \code{\link{precache}(x, ...)} trims the \verb{rcall://} prefix,
evaluates the remainder of the URI as R code,
and returns the resulting file path, for use in the \pkg{BiocFileCache}.
}
}

\author{
Thomas Sandmann
}
