% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEindex.R
\name{iSEEindex}
\alias{iSEEindex}
\title{iSEEindex app}
\usage{
iSEEindex(bfc, FUN.datasets)
}
\arguments{
\item{bfc}{An \code{\link[=BiocFileCache]{BiocFileCache()}} object.}

\item{FUN.datasets}{A function that returns a \code{data.frame} of metadata for available data sets.}
}
\value{
An \code{\link[=iSEE]{iSEE()}} app with a custom landing page using a \code{\link[=BiocFileCache]{BiocFileCache()}} to cache a selection of data sets.
}
\description{
iSEEindex app
}
\examples{
library(BiocFileCache)
bfc <- BiocFileCache()
dataset_fun <- function() {
    out <- read.csv(system.file(package = "iSEEindex", "datasets.csv"))
}

app <- iSEEindex(bfc, dataset_fun)

if (interactive()) {
  shiny::runApp(app, port = 1234)
}
}
