% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEindex.R
\name{iSEEindex}
\alias{iSEEindex}
\title{iSEEindex App}
\usage{
iSEEindex(bfc, FUN.datasets, FUN.initial = NULL)
}
\arguments{
\item{bfc}{An \code{\link[=BiocFileCache]{BiocFileCache()}} object.}

\item{FUN.datasets}{A function that returns a \code{data.frame} of metadata for
available data sets.}

\item{FUN.initial}{A function that returns a \code{data.frame} of metadata for
available initial configuration states.}
}
\value{
An \code{\link[=iSEE]{iSEE()}} app with a custom landing page using a \code{\link[=BiocFileCache]{BiocFileCache()}} to cache a selection of data sets.
}
\description{
Generate an \pkg{iSEE} app that includes a landing page enabling
users to choose from a custom set of data sets and initial configuration
states prepared by the app maintainer.
}
\section{Data Sets Metadata}{

The function passed to the argument \code{FUN.datasets} must return a \code{data.frame} that contains the following columns:

\describe{
\item{id}{A unique identifier for the data set.}
\item{label}{A short human-readable title for the data set, displayed in the 'Info' panel when the data set is selected.}
\item{description}{A more detailed description of the data set, displayed in the 'Info' panel when the data set is selected.}
\item{uri}{A Uniform Resource Identifier (URI) that indicates the location of the data file that contains the data set.}
}
}

\examples{
library(BiocFileCache)
bfc <- BiocFileCache(cache = tempdir())

dataset_fun <- function() {
    read.csv(system.file(package="iSEEindex", "datasets.csv"))
}

initial_fun <- function() {
    read.csv(system.file(package = "iSEEindex", "initial.csv"))
}

app <- iSEEindex(bfc, dataset_fun, initial_fun)

if (interactive()) {
  shiny::runApp(app, port = 1234)
}
}
\author{
Kevin Rue-Albrecht
}
