% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEindexResource-class.R
\name{iSEEindexHttpsResource-class}
\alias{iSEEindexHttpsResource-class}
\alias{precache,iSEEindexHttpsResource-method}
\title{The iSEEindexHttpsResource class}
\description{
The iSEEindexHttpsResource class represents a resource accessible through
an HTTPS link.
}
\section{Slot overview}{

This class inherits all slots from its parent class \linkS4class{iSEEindexResource}.
}

\section{Supported methods}{

In the following code snippets, \code{x} is an instance of a \linkS4class{iSEEindexHttpsResource} class.
Refer to the documentation for each method for more details on the remaining arguments.

\itemize{
\item \code{\link{precache}(x, ,..)} returns the URI to the resource as-is,
because The \pkg{BiocFileCache} naturally supports the HTTPS protocol.
}
}

\author{
Kevin Rue-Albrecht
}
