% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEindexResource-class.R
\name{iSEEindexRcallResource-class}
\alias{iSEEindexRcallResource-class}
\alias{precache,iSEEindexRcallResource-method}
\title{The iSEEindexRcallResource class}
\description{
The iSEEindexRcallResource class represents a resource accessible through
the result of an R call.
A URI for this type of resource uses the prefix \dQuote{rcall://}.
}
\details{
The URI must contain valid R code, once the prefix \verb{rcall://} is removed.
The code must return the file path to an existing R script
that produces an object called \code{initial} that contains a valid configuration
for an \pkg{iSEE} app.

For instance:

\if{html}{\out{<div class="sourceCode">}}\preformatted{rcall://system.file(package='iSEEindex','ReprocessedAllenData_config_01.R')
}\if{html}{\out{</div>}}
}
\section{Slot overview}{

This class inherits all slots from its parent class \linkS4class{iSEEindexResource}.
}

\section{Supported methods}{

In the following code snippets, \code{x} is an instance of a \linkS4class{iSEEindexRcallResource} class.
Refer to the documentation for each method for more details on the remaining arguments.

\itemize{
\item \code{\link{precache}(x, ...)} trims the \verb{rcall://} prefix,
evaluates the remainder of the URI as R code, and caches a copy of the
resource located at the resulting file path using \pkg{BiocFileCache},
before returning the file path to the cached file.
}
}

\author{
Kevin Rue-Albrecht
}
