% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEindexResource-class.R
\name{iSEEindexResource-class}
\alias{iSEEindexResource-class}
\alias{show,iSEEindexResource-method}
\alias{precache,iSEEindexResource-method}
\title{The iSEEindexResource class}
\description{
The iSEEindexResource class represents an undefined resource.
}
\section{Slot overview}{

\itemize{
\item \code{uri}, a character scalar specifying the URI to a resource.
}
}

\section{Supported methods}{

In the following code snippets, \code{x} is an instance of a \linkS4class{iSEEindexResource} class.
Refer to the documentation for each method for more details on the remaining arguments.

\itemize{
\item \code{\link{precache}(x, ,..)} throws an error, encouraging users to develop a method for derived classes that are not suported yet.
}
}

\author{
Kevin Rue-Albrecht
}
