% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEindexResource-class.R
\name{iSEEindexResource-class}
\alias{iSEEindexResource-class}
\alias{precache,iSEEindexResource-method}
\alias{show,iSEEindexResource-method}
\title{The iSEEindexResource class}
\usage{
\S4method{show}{iSEEindexResource}(object)

\S4method{precache}{iSEEindexResource}(x, bfc, id, ...)
}
\arguments{
\item{object}{An \code{iSEEindexResource} object.}

\item{x}{An \code{\linkS4class{iSEEindexResource}} object.}

\item{bfc}{A \code{\link[=BiocFileCache]{BiocFileCache()}} object.}

\item{id}{A data set identifier as a character scalar.}

\item{...}{additional arguments passed to and from other methods.}
}
\value{
\code{show()} returns \code{NULL} after displaying a summary of the object.

\code{precache()} throws an error if no method is found for the derived class.
}
\description{
The iSEEindexResource class is a virtual class from which classes of
supported resource must be derived.
}
\section{Slot overview}{

\itemize{
\item \code{uri}, a character scalar specifying the URI of a resource.
}
}

\section{Supported methods}{

In the following code snippets, \code{x} is an instance of a
\code{\linkS4class{iSEEindexResource}} class.
Refer to the documentation for each method for more details on the remaining
arguments.

\itemize{
\item \code{\link{precache}(x, bfc, id, ...)} throws an error, encouraging
users to develop a method for derived classes that are not supported yet.
}
}

\examples{
showClass("iSEEindexResource")
}
\author{
Kevin Rue-Albrecht
}
