% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEindexResource-class.R
\docType{methods}
\name{iSEEindexResource-generics}
\alias{iSEEindexResource-generics}
\alias{precache}
\title{Download and Cache Resources}
\description{
An overview of the generics for downloading and caching resources.
}
\section{Preparing and caching resources}{

\code{precache(x, bfc, id, ...)} potentially downloads a resource from an URI, caches it, and returns the path to the cached file:
\itemize{
\item \code{x}, a \code{\linkS4class{iSEEindexResource}} object.
\item \code{bfc}, a \code{\link[=BiocFileCache]{BiocFileCache()}} object.
\item \code{id}, a data set identifier as a character scalar..
\item \code{...}, additional arguments passed to and from other methods.
}
}

\examples{
library(BiocFileCache)
bfc <- BiocFileCache(cache = tempdir())

x <- iSEEindexRcallResource(list(
  uri = "rcall://system.file(package='iSEEindex','ReprocessedAllenData_config_01.R')"
))
precache(x, bfc, "ID0")
}
\author{
Kevin Rue-Albrecht
}
